Initial 2D cutting complete
